+++ paper/paper.bib
@@ -19,7 +19,7 @@ volume = {172},
 pages = {107148},
 year = {2020},
 issn = {1389-1286},
-doi = {https://doi.org/10.1016/j.comnet.2020.107148},
+doi = {10.1016/j.comnet.2020.107148},
 url = {https://www.sciencedirect.com/science/article/pii/S138912862030116X},
 author = {Panagiotis Radoglou-Grammatikis and Panagiotis Sarigiannidis and Thomas Lagkas and Ioannis Moscholios},
 keywords = {Precision agriculture (PA), Remote sensing (RS), Unmanned aerial vehicle (UAV)},
@@ -27,14 +27,17 @@ abstract = {Climate change has introduced significant challenges that can affect
 }
 
 @article{xue2017,
-  title={Significant remote sensing vegetation indices: A review of developments and applications},
-  author={Xue, Jinru and Su, Baofeng},
-  journal={Journal of sensors},
-  volume={2017},
-  number={1},
-  pages={1353691},
-  year={2017},
-  publisher={Wiley Online Library}
+author = {Xue, Jinru and Su, Baofeng},
+title = {Significant Remote Sensing Vegetation Indices: A Review of Developments and Applications},
+journal = {Journal of Sensors},
+volume = {2017},
+number = {1},
+pages = {1353691},
+doi = {10.1155/2017/1353691},
+url = {https://onlinelibrary.wiley.com/doi/abs/10.1155/2017/1353691},
+eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1155/2017/1353691},
+abstract = {Vegetation Indices (VIs) obtained from remote sensing based canopies are quite simple and effective algorithms for quantitative and qualitative evaluations of vegetation cover, vigor, and growth dynamics, among other applications. These indices have been widely implemented within RS applications using different airborne and satellite platforms with recent advances using Unmanned Aerial Vehicles (UAV). Up to date, there is no unified mathematical expression that defines all VIs due to the complexity of different light spectra combinations, instrumentation, platforms, and resolutions used. Therefore, customized algorithms have been developed and tested against a variety of applications according to specific mathematical expressions that combine visible light radiation, mainly green spectra region, from vegetation, and nonvisible spectra to obtain proxy quantifications of the vegetation surface. In the real-world applications, optimization VIs are usually tailored to the specific application requirements coupled with appropriate validation tools and methodologies in the ground. The present study introduces the spectral characteristics of vegetation and summarizes the development of VIs and the advantages and disadvantages from different indices developed. This paper reviews more than 100 VIs, discussing their specific applicability and representativeness according to the vegetation of interest, environment, and implementation precision. Predictably, research, and development of VIs, which are based on hyperspectral and UAV platforms, would have a wide applicability in different areas.},
+year = {2017}
 }
 
 @article{PANG2020,
@@ -44,7 +47,7 @@ volume = {178},
 pages = {105766},
 year = {2020},
 issn = {0168-1699},
-doi = {https://doi.org/10.1016/j.compag.2020.105766},
+doi = {10.1016/j.compag.2020.105766},
 url = {https://www.sciencedirect.com/science/article/pii/S0168169920311376},
 author = {Yan Pang and Yeyin Shi and Shancheng Gao and Feng Jiang and Arun-Narenthiran Veeranampalayam-Sivakumar and Laura Thompson and Joe Luck and Chao Liu},
 keywords = {Plant population, Deep learning, RCNN, Remote sensing, UAS},
@@ -81,7 +84,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="234--241",
[…]
