### 2025-05-01 #8124 @egemose
Hi! I have updated the paper with missing DOIs and fixed the ones with invalid DOIs.
