### 2025-05-01 #8124 @egemose
I have no suggestions for potential reviewers. Reviewers in the field of computer vision seems most relevant, but knowledge of rasters and GIS is also preferred.
