### 2025-05-19 #8124 @ujjwalkarn
@ggsharma, @pchlap, @pr4deepr -- would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/review_criteria.htm](https://joss.readthedocs.io/en/latest/review_criteria.html)
