### 2025-05-28 #8124 @tuelwer
> [@sashakolpakov](https://github.com/sashakolpakov) [@Aljanabi1989](https://github.com/Aljanabi1989) [@tuelwer](https://github.com/tuelwer) -- would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [https://joss.readthedocs.io/en/latest/review_criteria.htm](https://joss.readthedocs.io/en/latest/review_criteria.html)

Yes!
