### 2025-05-28 #8297 @editorialbot
[REVIEW]: CDC: a Python package for Color Distance Calculations
**Submitting author:** @egemose (<a href="http://orcid.org/0000-0002-6765-8216">Henrik Dyrberg Egemose</a>)
**Repository:** https://github.com/henrikmidtiby/CDC
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.2
