### 2025-06-12 #8297 @ujjwalkarn
👋 @egemose - thank you for your submission to JOSS. I am going to run this through review with our larger editorial board to ensure it meets our requirements for substantial scholarly effort and will be back in touch as soon as possible.
